chyba działa

+++ app/src/main/java/com/example/note/Model/AlarmItem.kt
@@ -5,5 +5,5 @@ import java.time.LocalDateTime
 data class AlarmItem(
     val time: LocalDateTime,
     val message: String,
-    val id: Int?
+    val id: Long?
 )
